requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
discord==2.3.2
discord.py==2.6.3
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
typing_extensions==4.13.2
yarl==1.15.2
numpy>=1.26
faiss-cpu>=1.7.4
requests>=2.32
pydantic>=2.8
tqdm>=4.66
